JAV-98 moved integration tests without docker into one parent module
