+++ integration-tests/jaxrs-tests/pom.xml
@@ -19,13 +19,13 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>demo-springmvc</artifactId>
-    <groupId>io.servicecomb.demo</groupId>
+    <artifactId>integration-tests</artifactId>
+    <groupId>io.servicecomb.tests</groupId>
     <version>0.1.0-m3-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>springmvc-tests</artifactId>
+  <artifactId>jaxrs-tests</artifactId>
 
   <dependencies>
     <dependency>
@@ -34,7 +34,7 @@
     </dependency>
     <dependency>
       <groupId>io.servicecomb</groupId>
-      <artifactId>provider-springmvc</artifactId>
+      <artifactId>provider-jaxrs</artifactId>
     </dependency>
   </dependencies>
