+++ integration-tests/jaxrs-tests/src/test/resources/microservice.yaml
