+++ integration-tests/springmvc-tests/src/test/java/io/servicecomb/demo/springmvc/tests/FrenchGreetingRestEndpoint.java
