+++ integration-tests/springmvc-tests/src/test/resources/microservice.yaml
