+++ parent/pom.xml
@@ -77,7 +77,7 @@
 					<version>4.3.0</version>
 					<configuration>
 						<jacocoReports>
-							<jacocoReport>${project.basedir}/integration-tests/target/site/jacoco-aggregate/jacoco.xml</jacocoReport>
+							<jacocoReport>${project.basedir}/coverage-reports/target/site/jacoco-aggregate/jacoco.xml</jacocoReport>
 						</jacocoReports>
 					</configuration>
 				</plugin>
